requirements.txt
flask
flask-login
flask-sqlalchemy
flask-session
python-dotenv
langchain
openai
sqlalchemy
werkzeug
gunicorn
tiktoken
flask
GitPython
python-dotenv
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
langchain_experimental
langchain-google-genai
tiktoken
faiss-cpu
protobuf
sentence-transformers
langchain_chroma